# backend/openapi.schema.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "NinjaAPI",
    "version": "1.0.0",
    "description": ""
  },
  "paths": {
    "/api/actors/{actor_id}": {
      "get": {
        "operationId": "videocreator_routers_actors_get_actor",
        "summary": "Get Actor",
        "parameters": [
          {
            "in": "path",
            "name": "actor_id",
            "schema": {
              "title": "Actor Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Get an actor\n    by reading yamls/cast/actors directory\n    the id is the filename",
        "tags": [
          "actors"
        ],
        "security": [
          {
            "GlobalAuth": []
          }
        ]
      }
    },
    "/api/actors/": {
      "get": {
        "operationId": "videocreator_routers_actors_get_actors",
        "summary": "Get Actors",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Get all actors\nby reading yamls/cast/actors directory",
        "tags": [
          "actors"
        ],
        "security": [
          {
            "GlobalAuth": []
          }
        ]
      }
    },
    "/api/programs/{program_id}": {
      "get": {
        "operationId": "videocreator_routers_programs_get_program",
        "summary": "Get Program",
        "parameters": [
          {
            "in": "path",
            "name": "program_id",
            "schema": {
              "title": "Program Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Get a program\nby reading yamls/programs directory\nthe id is the filename",
        "tags": [
          "programs"
        ],
        "security": [
          {
            "GlobalAuth": []
          }
        ]
      }
    },
    "/api/programs/": {
      "get": {
        "operationId": "videocreator_routers_programs_get_programs",
        "summary": "Get Programs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Get all programs\nby reading yamls/programs directory",
        "tags": [
          "programs"
        ],
        "security": [
          {
            "GlobalAuth": []
          }
        ]
      }
    },
    "/api/projects/": {
      "get": {
        "operationId": "videocreator_routers_projects_get_projects",
        "summary": "Get Projects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Get all projects\nby reading output/projects directory",
        "tags": [
          "projects"
        ],
        "security": [
          {
            "GlobalAuth": []
          }
        ]
      },
      "post": {
        "operationId": "videocreator_routers_projects_create_project",
        "summary": "Create Project",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Create a project\ncreate a directory output/projects/{project_id}",
        "tags": [
          "projects"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AICPProject"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "GlobalAuth": []
          }
        ]
      }
    },
    "/api/projects/{project_id}": {
      "get": {
        "operationId": "videocreator_routers_projects_get_project",
        "summary": "Get Project",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "schema": {
              "title": "Project Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Get a project\nby reading output/projects/{project_id}",
        "tags": [
          "projects"
        ],
        "security": [
          {
            "GlobalAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AICPActor": {
        "description": "Class encapsulating the actor for an AICP video.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "catch_phrase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catch Phrase"
          },
          "physical_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Physical Description"
          },
          "speaker": {
            "title": "Speaker",
            "type": "string"
          },
          "speaker_text_temp": {
            "default": 0.7,
            "title": "Speaker Text Temp",
            "type": "number"
          },
          "speaker_waveform_temp": {
            "default": 0.7,
            "title": "Speaker Waveform Temp",
            "type": "number"
          }
        },
        "required": [
          "name",
          "bio",
          "catch_phrase",
          "physical_description",
          "speaker"
        ],
        "title": "AICPActor",
        "type": "object"
      },
      "AICPProductionConfig": {
        "description": "Class encapsulating the production config for an AICP video.",
        "properties": {
          "video_width": {
            "default": 1920,
            "title": "Video Width",
            "type": "integer"
          },
          "video_height": {
            "default": 1080,
            "title": "Video Height",
            "type": "integer"
          },
          "sd_base_image_width": {
            "default": 768,
            "title": "Sd Base Image Width",
            "type": "integer"
          },
          "sd_base_image_height": {
            "default": 432,
            "title": "Sd Base Image Height",
            "type": "integer"
          },
          "enable_subtitles": {
            "default": false,
            "title": "Enable Subtitles",
            "type": "boolean"
          },
          "voiceline_synced_storyboard": {
            "default": false,
            "title": "Voiceline Synced Storyboard",
            "type": "boolean"
          },
          "num_images_per_scene": {
            "default": 1,
            "title": "Num Images Per Scene",
            "type": "integer"
          }
        },
        "title": "AICPProductionConfig",
        "type": "object"
      },
      "AICPProgram": {
        "description": "Class encapsulating the program for an AICP video.",
        "properties": {
          "title": {
            "title": "Title",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "prompt_placeholder_text": {
            "title": "Prompt Placeholder Text",
            "type": "string"
          },
          "script_rules": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Script Rules"
          },
          "storyboard_rules": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storyboard Rules"
          },
          "music_rules": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Music Rules"
          }
        },
        "required": [
          "title",
          "description",
          "prompt_placeholder_text"
        ],
        "title": "AICPProgram",
        "type": "object"
      },
      "AICPProject": {
        "description": "Class encapsulating the project for an AICP video.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "program": {
            "$ref": "#/components/schemas/AICPProgram"
          },
          "production_config": {
            "$ref": "#/components/schemas/AICPProductionConfig"
          },
          "actors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AICPActor"
            },
            "title": "Actors",
            "type": "array"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "program",
          "production_config",
          "seed"
        ],
        "title": "AICPProject",
        "type": "object"
      }
    },
    "securitySchemes": {
      "GlobalAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": []
}